lint: fix lints
