+++ packages/auth-ui/src/components/ui/badge.tsx
@@ -33,4 +33,4 @@ function Badge({
   return <Comp data-slot="badge" className={cn(badgeVariants({ variant }), className)} {...props} />;
 }
 
-export { Badge, badgeVariants };
+export { Badge };
