+++ packages/auth-ui/src/fetch/client.ts
@@ -92,3 +92,5 @@ export const updateApiBaseUrl = (baseUrl: string) => {
 };
 
 export const siteApis = await createSiteApis();
+
+export const getAvailableSites = () => Object.keys(siteApis || {});
